EDGECLOUD-641 clusterinst deployment
