+++ controller/upgrade_testfiles/AddClusterInstDeploymentField_post.etcd
@@ -1,0 +1,4 @@
+1/ClusterInst/{"cluster_key":{"name":"SmallCluster"},"cloudlet_key":{"operator_key":{"name":"tmus"},"name":"tmus-cloud-1"},"developer":"AcmeAppCo"}
+{"key":{"cluster_key":{"name":"SmallCluster"},"cloudlet_key":{"operator_key":{"name":"tmus"},"name":"tmus-cloud-1"},"developer":"AcmeAppCo"},"flavor":{"name":"x1.small"},"liveness":"LivenessStatic","state":"Ready","ip_access":"IpAccessShared","node_flavor":"flavor1","deployment":"kubernetes","num_masters":1,"num_nodes":2}
+1/ClusterInst/{"cluster_key":{"name":"SmallCluster"},"cloudlet_key":{"operator_key":{"name":"tmus"},"name":"tmus-cloud-2"},"developer":"AcmeAppCo"}
+{"key":{"cluster_key":{"name":"SmallCluster"},"cloudlet_key":{"operator_key":{"name":"tmus"},"name":"tmus-cloud-2"},"developer":"AcmeAppCo"},"flavor":{"name":"x1.small"},"liveness":"LivenessStatic","state":"Ready","ip_access":"IpAccessDedicated","allocated_ip":"dynamic","node_flavor":"flavor1","deployment":"kubernetes","num_masters":1,"num_nodes":2}
