+++ edgeproto/clusterinst.proto
@@ -51,7 +51,9 @@ message ClusterInst {
   string allocated_ip = 8 [(protogen.backend) = true, (protogen.hidetag) = "nocmp"];
   // Cloudlet specific node flavor
   string node_flavor = 11 [(protogen.backend) = true,(protogen.hidetag) = "nocmp"];
-    // number of masters
+  // Deployment target (kubernetes, docker, kvm, etc)
+  string deployment = 15 [(protogen.backend) = true];
+  // number of masters
   uint32 num_masters = 13;
   // number of nodes 
   uint32 num_nodes = 14;
