+++ edgeproto/version.proto
@@ -14,6 +14,7 @@ enum VersionHash {
   HASH_52e6980599cd59bbbd0de8d5f4d53d4b = 2;
   HASH_00bdcfa956ca4ee42be87abcd8fcaf1c = 3 [(protogen.upgrade_func) = "AddClusterInstKeyToAppInstKey"];
   HASH_0d2d9c0b07ad989e96fb3b3a44924316 = 4 [(protogen.upgrade_func) = "LintFixAppInstFields"];
+  HASH_2b79f0b6e402045ee5f68d697b9386ae = 5 [(protogen.upgrade_func) = "AddClusterInstDeploymentField"];
   option (protogen.version_hash) = true;
-  option (protogen.version_hash_salt) = "1";
+  option (protogen.version_hash_salt) = "2";
 }
