+++ gencmd/clusterinst.cmd.go
@@ -85,7 +85,7 @@ func ClusterInstKeyWriteOutputOne(obj *edgeproto.ClusterInstKey) {
 	}
 }
 func ClusterInstSlicer(in *edgeproto.ClusterInst) []string {
-	s := make([]string, 0, 13)
+	s := make([]string, 0, 14)
 	if in.Fields == nil {
 		in.Fields = make([]string, 1)
 	}
@@ -106,13 +106,14 @@ func ClusterInstSlicer(in *edgeproto.ClusterInst) []string {
 	s = append(s, edgeproto.IpAccess_CamelName[int32(in.IpAccess)])
 	s = append(s, in.AllocatedIp)
 	s = append(s, in.NodeFlavor)
+	s = append(s, in.Deployment)
 	s = append(s, strconv.FormatUint(uint64(in.NumMasters), 10))
 	s = append(s, strconv.FormatUint(uint64(in.NumNodes), 10))
 	return s
 }
 
 func ClusterInstHeaderSlicer() []string {
-	s := make([]string, 0, 13)
+	s := make([]string, 0, 14)
 	s = append(s, "Fields")
 	s = append(s, "Key-ClusterKey-Name")
 	s = append(s, "Key-CloudletKey-OperatorKey-Name")
@@ -127,6 +128,7 @@ func ClusterInstHeaderSlicer() []string {
 	s = append(s, "IpAccess")
 	s = append(s, "AllocatedIp")
 	s = append(s, "NodeFlavor")
+	s = append(s, "Deployment")
 	s = append(s, "NumMasters")
 	s = append(s, "NumNodes")
 	return s
@@ -555,6 +557,7 @@ func init() {
 	ClusterInstFlagSet.StringVar(&ClusterInstInIpAccess, "ipaccess", "", "one of [IpAccessUnknown IpAccessDedicated IpAccessDedicatedOrShared IpAccessShared]")
 	ClusterInstNoConfigFlagSet.StringVar(&ClusterInstIn.AllocatedIp, "allocatedip", "", "AllocatedIp")
 	ClusterInstNoConfigFlagSet.StringVar(&ClusterInstIn.NodeFlavor, "nodeflavor", "", "NodeFlavor")
+	ClusterInstFlagSet.StringVar(&ClusterInstIn.Deployment, "deployment", "", "Deployment")
 	ClusterInstFlagSet.Uint32Var(&ClusterInstIn.NumMasters, "nummasters", 0, "NumMasters")
 	ClusterInstFlagSet.Uint32Var(&ClusterInstIn.NumNodes, "numnodes", 0, "NumNodes")
 	ClusterInstInfoFlagSet.StringVar(&ClusterInstInfoIn.Key.ClusterKey.Name, "key-clusterkey-name", "", "Key.ClusterKey.Name")
@@ -618,6 +621,9 @@ func ClusterInstSetFields() {
 	}
 	if ClusterInstNoConfigFlagSet.Lookup("nodeflavor").Changed {
 		ClusterInstIn.Fields = append(ClusterInstIn.Fields, "11")
+	}
+	if ClusterInstFlagSet.Lookup("deployment").Changed {
+		ClusterInstIn.Fields = append(ClusterInstIn.Fields, "15")
 	}
 	if ClusterInstFlagSet.Lookup("nummasters").Changed {
 		ClusterInstIn.Fields = append(ClusterInstIn.Fields, "13")
