+++ gencmd/version.cmd.go
@@ -22,6 +22,7 @@ var VersionHashStrings = []string{
 	"Hash52E6980599Cd59Bbbd0De8D5F4D53D4B",
 	"Hash00Bdcfa956Ca4Ee42Be87Abcd8Fcaf1C",
 	"Hash0D2D9C0B07Ad989E96Fb3B3A44924316",
+	"Hash2B79F0B6E402045Ee5F68D697B9386Ae",
 }
 
 func init() {
