+++ setup-env/e2e-tests/data/appdata_clusterInst1_cloudlet1_appInst0.yml
@@ -34,6 +34,7 @@ clusterinsts:
     name: x1.small
   liveness: LivenessStatic
   ipaccess: IpAccessShared
+  deployment: kubernetes
   nummasters: 1
   numnodes: 1
