+++ setup-env/e2e-tests/data/appdata_clusterInst1_cloudlet1_appInst1_show.yml
@@ -46,6 +46,7 @@ clusterinsts:
     name: x1.small
   liveness: LivenessStatic
   ipaccess: IpAccessShared
+  deployment: kubernetes
   nummasters: 1
   numnodes: 1
 developers:
