+++ setup-env/e2e-tests/data/appdata_no_appinst_show.yml
@@ -35,6 +35,7 @@ clusterinsts:
     name: x1.small
   liveness: LivenessStatic
   ipaccess: IpAccessDedicated
+  deployment: kubernetes
   nummasters: 1
   numnodes: 1
 - key:
@@ -49,6 +50,7 @@ clusterinsts:
     name: x1.small
   liveness: LivenessStatic
   ipaccess: IpAccessShared
+  deployment: kubernetes
   nummasters: 1
   numnodes: 1
