+++ setup-env/e2e-tests/data/appdata_show.yml
@@ -34,6 +34,7 @@ clusterinsts:
     name: x1.small
   liveness: LivenessStatic
   ipaccess: IpAccessShared
+  deployment: kubernetes
   nummasters: 1
   numnodes: 2
 - key:
@@ -48,6 +49,7 @@ clusterinsts:
     name: x1.small
   liveness: LivenessStatic
   ipaccess: IpAccessDedicated
+  deployment: kubernetes
   nummasters: 1
   numnodes: 2
