+++ setup-env/e2e-tests/data/show2.yml
@@ -31,6 +31,7 @@ clusterinsts:
   liveness: LivenessDynamic
   auto: true
   ipaccess: IpAccessShared
+  deployment: kubernetes
   nummasters: 1
   numnodes: 1
 - key:
@@ -46,6 +47,7 @@ clusterinsts:
   liveness: LivenessDynamic
   auto: true
   ipaccess: IpAccessShared
+  deployment: kubernetes
   nummasters: 1
   numnodes: 1
